Fix Y2K38: Change timer and timestamp types from 32-bit to 64-bit
This commit addresses the Y2K38 (2038 timestamp overflow) issue by changing
all timer and timestamp fields from 32-bit to 64-bit types:

Header changes:
- stp_timer.h: Changed TIMER.value from UINT32 to UINT64, updated function
  signatures for start_timer, timer_expired, get_timer_value, sys_get_seconds
- stp.h: Changed topology_change_tick, last_expiry_time, last_bpdu_rx_time
  from UINT32 to UINT64
- stp_dbsync.h: Changed topology_change_time and root_protect_timer from
  uint32_t to uint64_t
- mstp.h: Changed last_expiry_time from UINT32 to UINT64

Implementation changes:
- stp_timer.c: Updated sys_get_seconds() return type and all timer functions
- stp.c: Updated local variables in hello_timer_expiry()
- stp_util.c: Updated local variables in stputil_process_bpdu()

Added tests/test_y2k38.cpp with comprehensive tests verifying:
- TIMER struct value field is 64-bit
- All timestamp fields can store values beyond 2038
- Timer arithmetic works correctly for post-Y2K38 values

+++ include/stp_dbsync.h
@@ -39,7 +39,8 @@ typedef struct {
     uint8_t forward_delay;
     uint8_t hold_time;
 
-    uint32_t topology_change_time; // time of last tc event
+    /* Y2K38 Fix: Changed from uint32_t to uint64_t for timestamp safety */
+    uint64_t topology_change_time; // time of last tc event
     uint32_t topology_change_count;
 
     char root_bridge_id[STP_SYNC_BRIDGE_ID_LEN];
@@ -70,7 +71,8 @@ typedef struct {
     uint32_t rx_config_bpdu;
     uint32_t tx_tcn_bpdu;
     uint32_t rx_tcn_bpdu;
-    uint32_t root_protect_timer;
+    /* Y2K38 Fix: Changed from uint32_t to uint64_t for timestamp safety */
+    uint64_t root_protect_timer;
 
     uint8_t clear_stats;
     uint32_t modified_fields;
